Refactoring: Extract FielValueMirror into separate file
