+++ BTCPayServer/Forms/HtmlInputFormProvider.cs
@@ -5,27 +5,6 @@ using BTCPayServer.Validation;
 
 namespace BTCPayServer.Forms;
 
-public class FieldValueMirror : IFormComponentProvider
-{
-    public string View { get; } = null;
-    public void Validate(Form form, Field field)
-    {
-        if (form.GetFieldByFullName(field.Value) is null)
-        {
-            field.ValidationErrors = new List<string>() { $"{field.Name} requires {field.Value} to be present" };
-        }
-    }
-
-    public void Register(Dictionary<string, IFormComponentProvider> typeToComponentProvider)
-    {
-        typeToComponentProvider.Add("mirror", this);
-    }
-
-    public string GetValue(Form form, Field field)
-    {
-        return form.GetFieldByFullName(field.Value)?.Value;
-    }
-}
 public class HtmlInputFormProvider : FormComponentProviderBase
 {
     public override void Register(Dictionary<string, IFormComponentProvider> typeToComponentProvider)
